move indexvarrelgraph to new file
